rename slider file and add search to getproducts path
