+++ components/layout/product-grid-items.tsx
@@ -11,7 +11,7 @@ export default function ProductGridItems({
   return (
     <>
       {products.map((product) => (
-        <Grid.Item key={product.handle} className="animate-fadeIn">
+        <Grid.Item key={product.id} className="animate-fadeIn">
           <Link
             className="relative inline-block h-full w-full"
             href={`/product/${product.handle}`}
