+++ components/slider.tsx
